chore: support not show the native sql
